Update login, delete user, etc.

+++ controllers/deleteUserController.js
@@ -4,6 +4,10 @@ module.exports = async (req, res) => {
   // /users/:userId, DELETE
   let userId = req.params.userId;
   if (!userId) res.sendStatus(400);
+  if (userId !== req.session.userId) {
+    console.log("Danger: User trying to delete another user");
+    return res.redirect("/");
+  }
   try {
     const [result] = await pool.query(
       `DELETE FROM users WHERE id = ?`,
